Move Matched out of matchers._base into its own file to resolve docs cross referencing issues
