+++ src/traccuracy/metrics/_divisions.py
@@ -41,14 +41,14 @@ from typing import TYPE_CHECKING
 import numpy as np
 
 from traccuracy._tracking_graph import NodeFlag
-from traccuracy.matchers._base import Matched
+from traccuracy.matchers._matched import Matched
 from traccuracy.track_errors._divisions import evaluate_division_events
 
 from ._base import Metric
 
 if TYPE_CHECKING:
     from traccuracy import TrackingGraph
-    from traccuracy.matchers import Matched
+    from traccuracy.matchers._matched import Matched
 
 logger = logging.getLogger(__name__)
